// sql-voice-query/src/api-router/raindrop.gen.ts
// Code generated by raindrop build, DO NOT EDIT.
import { ActorNamespace, Ai, Annotation, App, Logger, MRNObject, ServiceStub, SmartMemory, SmartSql, Tracer,  } from '@liquidmetal-ai/raindrop-framework';

export interface Env {
  _raindrop: {
    app: App;
  };
  AI: Ai;
  annotation: Annotation<Omit<MRNObject, 'type' | 'applicationName' | 'versionId'>>;
  CONVERSATION_MEMORY: SmartMemory;
  DATABASE_MANAGER: ServiceStub<import('../database-manager/index.js').default>;
  FINANCIAL_DATABASE: SmartSql;
  logger: Logger;
  tracer: Tracer;
  VOICE_PROCESSOR: ActorNamespace<import('../voice-processor/index.js').VoiceProcessor>;
}